Remove the old volume-based test setup for tet tallies
This setup involved compiling special checks into the tally code
and using a particular problem geometry to attempt to check that all
tally track length was being accounted for.  It has not been used
for some months at this point, and rather than leaving ugly
instrumentation in the code I have opted to remove the test entirely.

git-svn-id: file:///filespace/groups/cnerg/svn/DAG-MCNP/5.1.51/trunk/Test-meshtal@399 5ce35831-a721-4015-95db-45fe606c5212